[flang][hlfir] Only canonicalize forall_index if it can be erased
It seems the canonicalization was not correct: it cannot returned that
it failed if it did modify the IR.
This was revealed by a new MLIR sanity check added in
https://reviews.llvm.org/D144552.
I am not sure it is legit to return success if the operation being
canonicalized is not modified either. So only remove the loads if
they are the only uses of the forall_index.

Should fix (intermittent?) bot failures like
https://lab.llvm.org/buildbot/#/builders/179/builds/6251
since the new MLIR check was added.