[flang] Lower special bind(c) cases without binding labels
1. Deal with BIND(C,NAME="")

BIND(C,NAME="") is different from BIND(C). The latter implies that there
us a binding label which is the Fortran symbol name (no Fortran mangling
must be added like underscores). The former implies there is no binding
label (the name in the object file must be the same as if it there was
no BIND(C) attribute at all).

This is correctly implemented in the front-end, but lowering mistakenly
overrode this in the code dealing with the case where BIND(C) is
inherited from a procedure interface (which I am not sure can be handled
in name resolution SetBindNameOn, since it is not clear to me that the
procedure interface symbol is set already).

2. Deal with BIND(C) internal procedure

Also according to 18.10.2, BIND(C) does not give a p
Prevent name resolution from adding a label to them, otherwise,
bindc_internal_proc.f90 was not going through semantics (bogus error
about conflicting global names). Nothing TODO in lowering other than
removing the TODO.