[flang] Use mlir::DataLayout to resolve type size in codegen
Currently flang was generating "sizeof" GEP to compute type size and
leave LLVM resolve these expression to actual constant based on the
target data layout,

The ability to generate these constant expressions will soon be removed,
so the target data layout must be used in FIR to LLVM IR code generation
directly.

See https://github.com/llvm/llvm-project/issues/71507

Since LLVM IR dialect is being manipulated, an mlir::DataLayout must be
used to compute the mlir type size.  The mlir::DataLayout will only
contain the required if the moduleOp has an
mlir::DataLayoutSpecInterface with the data layout mapping for the
target.

Flang-new driver is already setting this attribute from the target info.

This patch modifies tco to set--up a default layout is none is already
set, which fits with the usage of tco as a target independent testing
tool for FIR transformation and codegen.

An option is also added to the FIR to LLVM codegen pass to set-up such
default layout if no layout was set, which allow target independent
testing with fir-opt tool. The option is off by default to avoid
production drivers and tools to use the default layout by accident.