[flang] Lower procedure pointer components
Lower procedure pointer components, except in the context of
structure constructor (left TODO).

Procedure pointer components lowering share most of the lowering
logic of procedure poionters with the following particularities:
- They are components, so an hlfir.designate must be generated to
  retrieve the procedure pointer address from its derived type base.
- They may have a PASS argument. While there is no dispatching as with
  type bound procedure, special care must be taken to retrieve the
  derived type component base in this case since semantics placed it
  in the argument list and not in the evaluate::ProcedureDesignator.

These components also bring a new level of recursive MLIR types since a
fir.type may now contain a component with an MLIR function type where
one of the argument is the fir.type itself. This required moving the
"derived type in construction" stackto the converter so that the object
and function type lowering utilities share the same state (currently the
function type utilty would end-up creating a new stack when lowering its
arguments, leading to infinite loops). The BoxedProcedurePass also
needed an update to deal with this recursive aspect.