[flang] Do not leave length parameters uninitialized  in descriptors
Descriptor addendum have a field to hold length parameters (currently only
one). This field is currently never used because flang does not lowered derived
types with length parameters.

However, leaving it uninitialized is causing bugs in code like gFTK where the
code is trying to sort POINTERs [1]. More precisely, it is an issue when two
pointers should compare equal (same base address), because the uninitialized
values in the addendum may differ depending on the "stack history" and
optimization level.

Always initialized the fields to zero.

[1]: https://github.com/Goddard-Fortran-Ecosystem/gFTL/blob/dc93a5fc2fe2b487590580d9f0a17d77ca3fe5c1/include/v1/templates/set_impl.inc#L312

The type being transferred to an integer array may look like:

   TYPE :: localwrapper
    TYPE(T), POINTER :: item
   END TYPE localwrapper

Which in flang case ends-up transferring a descriptor to an integer array.