[flang] Give internal linkage to internal functions
Internal function cannot be called directly from outside the host function, so
there is no point giving them external linkage. The only reason flang did is
because it is the default in MLIR.

Giving external linkage to them:
- prevents deleting them when not used/inlined by LLVM
- causes bugs with shared libraries (at least on linux x86-64) because
  the call to the internal function could lead to a dynamic loader call
  that would overwrite r10 (the static chain pointer) due to system calls
  and did not restore it since it seems it does not expect r10 to be used
  for PLT calls.

Note: the llvm.linkage attribute name cannot be obtained via a
getLinkageAttrName since it is not the same name as the one used in the LLVM
dialect. It is just a placeholder defined in
mlir/lib/Conversion/FuncToLLVM/FuncToLLVM.cpp until the func dialect gets a
real linkage model. So simply avoid hard coding it too many times in lowering.

f