[flang] Handle OPTIONAL polymorphic captured in internal procedure
The current code was doing an unconditional
`fir.store %optional_box to %host_link` which caused a crash when
%optional_box is absent because is is attempting to copy a descriptor
from a null address.

Add code to conditionally do the copy at runtime.
The polymorphic array case with lower bounds can be handled with the
array case that already deals with descriptor argument, just use that.