[flang] Do not hoist all scalar expressions from WHERE constructs
The HLFIR pass lowering WHERE (hlfir.where op) was too aggressive
in its hoisting of scalar sub-expressions from LHS/RHS/MASKS outside
of the loops generated for the WHERE construct.
This violated F'2023 10.2.3.2 point 10 that stipulated that elemental
operations must be evaluated only for elements corresponding to true
values, because scalar operations are still elemental, and hoisting
them is invalid if they could have side effects (e.g, division by zero)
and if the MASK is always false (i.e., the loop body is never evaluated).

The difficulty is that 10.2.3.2 point 9 mandates that nonelemental function
must be evaluated before the loops. So it is not possible to simply stop
hoisting non hlfir.elemental operations.
Marking calls with an elemental/nonelemental attribute would not allow
the pass to be correct if inlining is run before and drops this information,
beside, extracting the argument tree that may have been CSE-ed with the rest of
the expression evaluation would be a bit combursome.

Instead, lower nonelemental calls into a new hlfir.exactly_once operation
that will allow retaining the information that the operations contained inside
its region must be hoisted. This allows inlining to operate before if desired
in order to improve alias analysis.

The LowerHLFIROrderedAssignments pass is updated to only hoist the operations
contained inside hlfir.exactly_once bodies.